refactor: 숫자 추출 클래스와 사용자 입력 전처리 클래스 병합
