+++ src/main/java/calculator/Calculator.java
@@ -8,28 +8,24 @@ import java.util.List;
  */
 public class Calculator {
 
-    private final DigitExtractor digitExtractor;
-    private final OperationProcessor operationProcessor;
     private final InputProcessor inputProcessor;
+    private final OperationProcessor operationProcessor;
 
     /**
-     * Calculator 클래스의 생성자 - DigitExtractor, OperationProcessor, InputProcessor 객체 초기화
+     * Calculator 클래스의 생성자 - InputProcessor, OperationProcessor 객체 초기화
      */
     public Calculator() {
-        this.digitExtractor = new DigitExtractor();
-        this.operationProcessor = new OperationProcessor();
         this.inputProcessor = new InputProcessor();
+        this.operationProcessor = new OperationProcessor();
     }
 
     /**
      * 계산기 실행 - 사용자로부터 문자열을 입력 받아 연산을 수행하고 결과를 출력한다.
      */
     public void run() {
-        String operationInput = inputProcessor.getOperationInput(); // 연산 대상이 되는 문자열
-        String delimiterRegex = inputProcessor.getDelimiterRegex(); // 사용하는 구분자의 정규표현식
-        List<Integer> digits = digitExtractor.extractDigits(operationInput, delimiterRegex);
-        int result = operationProcessor.calculate(digits);
-        System.out.println("결과 : " + result);
+        List<Integer> digits = inputProcessor.extractDigits(); // 사용자 입력으로부터 덧셈의 피연산자(input)가 되는 숫자 추출
+        int result = operationProcessor.calculate(digits); // 덧셈 수행
+        System.out.println("결과 : " + result); // 계산 결과 출력
         Console.close();
     }
